Name every test file after the code it covers
`test_regressions_v090.py` was the only file in tests/ not named after a module.
It grouped tests by the defect they reproduce rather than by the code under
test, so its 37 cases were spread across ten modules and its three pcoa cases
sat on concerns that test_external.py already owned. Each case now lives in the
file for the module it exercises.

Also splits test_external.py per module, as tests/ already does for
preprocessing, so pcoa, umap, gseapy and goenrich each have their own file.
pcoa_utils goes with pcoa, the way test_tensor_helpers.py groups the small
tensor modules.

Test bodies, names, parametrizations and markers are carried over unchanged;
only the call style is adapted to the imports of each destination. The comment
blocks explaining each defect move with their tests, since they document why the
assertions exist.

One case is dropped rather than moved:
test_reorder_dimension_elements_without_metadata was a strict subset of
test_reorder_dimension_elements in test_plotting.py, which makes the same call
and additionally checks that the other dimensions keep their order. Its
explanation of the bug is folded into the test that stays.

Verified against pandas 3.0.5 and 2.3.3 on Python 3.12: 620 passing on both,
from 621, the difference being the dropped duplicate. The other 36 cases were
each confirmed to still be collected.
