+++ tests/test_analysis_tensor_downstream.py
@@ -7,6 +7,9 @@ import pandas as pd
 import pytest
 
 from cell2cell.analysis import tensor_downstream
+from cell2cell.analysis.tensor_downstream import (flatten_factor_ccc_networks,
+                                                 get_factor_specific_ccc_networks,
+                                                 get_lr_by_cell_pairs)
 
 
 LABELS = dict(sender_label='Sender Cells', receiver_label='Receiver Cells')
@@ -219,3 +222,129 @@ def test_get_lr_by_cell_pairs_rejects_unknown_labels(factorized_tensor):
         tensor_downstream.get_lr_by_cell_pairs(factorized_tensor, lr_label='Nope',
                                                sender_label='Sender Cells',
                                                receiver_label='Receiver Cells')
+
+
+# ---------------------------------------------------------------------------------
+# Cell-pair labels in flattened factor-specific networks
+#
+# `flatten_factor_ccc_networks` built the 'sender --> receiver' labels from the
+# SORTED cell names while flattening the values in the tensor's own order. Every
+# loading was assigned to the wrong cell pair when the tensor was not alphabetically
+# sorted, and `get_lr_by_cell_pairs` inherited the mislabeling.
+# ---------------------------------------------------------------------------------
+
+@pytest.mark.parametrize('orderby', ['senders', 'receivers'])
+def test_flatten_labels_match_values_on_unsorted_tensor(factorized_prebuilt_tensor, orderby):
+    '''The decisive regression test: 27/27 entries were mislabeled before the fix.'''
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    flat = flatten_factor_ccc_networks(networks, orderby=orderby)
+
+    # The fixture is deliberately unsorted, otherwise this test cannot fail
+    cells = list(factorized_prebuilt_tensor.order_names[2])
+    assert cells != sorted(cells)
+
+    for factor, network in networks.items():
+        for sender in network.index:
+            for receiver in network.columns:
+                label = '{} --> {}'.format(sender, receiver)
+                assert np.isclose(flat.loc[label, factor], network.loc[sender, receiver])
+
+
+def test_flatten_preserves_the_tensor_order(factorized_prebuilt_tensor):
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    flat = flatten_factor_ccc_networks(networks, orderby='senders')
+    cells = list(factorized_prebuilt_tensor.order_names[2])
+    expected = ['{} --> {}'.format(s, r) for s in cells for r in cells]
+    assert list(flat.index) == expected
+
+
+def test_flatten_groups_by_receiver_when_requested(factorized_prebuilt_tensor):
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    flat = flatten_factor_ccc_networks(networks, orderby='receivers')
+    cells = list(factorized_prebuilt_tensor.order_names[2])
+    expected = ['{} --> {}'.format(s, r) for r in cells for s in cells]
+    assert list(flat.index) == expected
+
+
+def test_flatten_realigns_networks_with_different_element_order(factorized_prebuilt_tensor):
+    '''Hand-assembled networks may not share an element order; values must still align.'''
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    factors = list(networks.keys())
+    shuffled = {factors[0]: networks[factors[0]],
+                factors[1]: networks[factors[1]].reindex(index=['C1', 'C2', 'C3'],
+                                                         columns=['C2', 'C3', 'C1'])}
+    flat = flatten_factor_ccc_networks(shuffled)
+    for factor in factors:
+        for sender in ['C1', 'C2', 'C3']:
+            for receiver in ['C1', 'C2', 'C3']:
+                label = '{} --> {}'.format(sender, receiver)
+                assert np.isclose(flat.loc[label, factor],
+                                  networks[factor].loc[sender, receiver])
+
+
+def test_flatten_keeps_only_common_elements(factorized_prebuilt_tensor):
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    factors = list(networks.keys())
+    subset = {factors[0]: networks[factors[0]],
+              factors[1]: networks[factors[1]].drop(index='C1', columns='C1')}
+    flat = flatten_factor_ccc_networks(subset)
+    assert len(flat) == 4                       # 2 senders x 2 receivers
+    assert not any('C1' in label for label in flat.index)
+
+
+def test_flatten_accepts_a_single_factor(factorized_prebuilt_tensor):
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    factor = list(networks.keys())[0]
+    flat = flatten_factor_ccc_networks({factor: networks[factor]})
+    assert list(flat.columns) == [factor]
+    for sender in networks[factor].index:
+        for receiver in networks[factor].columns:
+            label = '{} --> {}'.format(sender, receiver)
+            assert np.isclose(flat.loc[label, factor],
+                              networks[factor].loc[sender, receiver])
+
+
+def test_flatten_rejects_an_invalid_orderby(factorized_prebuilt_tensor):
+    networks = get_factor_specific_ccc_networks(factorized_prebuilt_tensor)
+    with pytest.raises(ValueError):
+        flatten_factor_ccc_networks(networks, orderby='not-an-option')
+
+
+def test_get_lr_by_cell_pairs_labels_match_values(factorized_prebuilt_tensor):
+    '''get_lr_by_cell_pairs inherited the mislabeling from the flattening.'''
+    tensor = factorized_prebuilt_tensor
+    networks = get_factor_specific_ccc_networks(tensor)
+    lr_loadings = tensor.factors['Ligand-Receptor Pairs']
+
+    result = get_lr_by_cell_pairs(tensor,
+                                  lr_label='Ligand-Receptor Pairs',
+                                  sender_label='Sender Cells',
+                                  receiver_label='Receiver Cells')
+
+    for cell_pair in result.columns:
+        sender, receiver = cell_pair.split(' --> ')
+        for lr_pair in result.index:
+            expected = sum(networks[f].loc[sender, receiver] * lr_loadings.loc[lr_pair, f]
+                           for f in networks)
+            assert np.isclose(result.loc[lr_pair, cell_pair], expected)
+
+
+# ---------------------------------------------------------------------------------
+# Natural ordering of factor names
+#
+# The factor names were sorted lexicographically, so decompositions with 10 or more
+# factors were returned as Factor 1, Factor 10, Factor 11, Factor 2, ...
+# ---------------------------------------------------------------------------------
+
+def test_get_lr_by_cell_pairs_accepts_other_factor_names(factorized_prebuilt_tensor):
+    '''Previously raised IndexError: the code did int(name.split(' ')[1]).'''
+    tensor = factorized_prebuilt_tensor
+    for key in tensor.factors:
+        tensor.factors[key].columns = ['component_1', 'component_10']
+
+    result = get_lr_by_cell_pairs(tensor,
+                                  lr_label='Ligand-Receptor Pairs',
+                                  sender_label='Sender Cells',
+                                  receiver_label='Receiver Cells')
+    assert result.shape[0] == len(tensor.order_names[1])
+    assert result.shape[1] == len(tensor.order_names[2]) ** 2
